Divided components into generic components and scenes
