+++ src/scenes/QR/styles.scss
